requirements.txt
# Основные зависимости для VPN приложения
Flask==2.3.3
Werkzeug==2.3.7

# База данных
psycopg2-binary==2.9.7

# Платежи
yookassa==3.0.0

# JWT токены
PyJWT==2.8.0

# Email
secure-smtplib==0.1.1

# Системные утилиты
psutil==5.9.6
python-dateutil==2.8.2

# QR коды
qrcode[pil]==7.4.2

# Дополнительные утилиты
requests==2.31.0
cryptography==41.0.7

# Telegram Bot
python-telegram-bot==20.7
